v0.4.0
changelog:
- glmnet_ts.py and test_glmnet_ts.py modules added
- model.py and test_model.py modules modified
